fix: 총 수익률 출력 포맷 수정

+++ src/main/java/lotto/controller/dto/output/WinningLottoResultDto.java
@@ -4,8 +4,8 @@ import java.util.Collections;
 import java.util.List;
 import lotto.common.constants.LottoRank;
 
-public record WinningLottoResultDto(List<LottoRank> winningResult, float profitRate) {
-    public WinningLottoResultDto(List<LottoRank> winningResult, float profitRate) {
+public record WinningLottoResultDto(List<LottoRank> winningResult, double profitRate) {
+    public WinningLottoResultDto(List<LottoRank> winningResult, double profitRate) {
         this.winningResult = Collections.unmodifiableList(winningResult);
         this.profitRate = profitRate;
     }
